rename installableIcon
