+++ src/mods/ModIcon.module.scss
