+++ src/mods/ModIcon.tsx
@@ -28,9 +28,9 @@ import {
   getPackageId,
   isBlueprint,
   isUnavailableRecipe,
-} from "@/utils/installableUtils";
+} from "@/utils/modUtils";
 import cx from "classnames";
-import styles from "./InstallableIcon.module.scss";
+import styles from "./ModIcon.module.scss";
 import { useGetMarketplaceListingsQuery } from "@/services/api";
 
 function getDefaultInstallableIcon(installable: Mod) {
